Bugfix: more directory separators
Windows pukes when slashes and backslashes are used together in an
"open" command in python. All the mkdir and file exists stuff seems to work
fine.

The first fix I submitted was just for the one error I got while testing ink.
As I've continued to play with it, I came across another place where the
issue threw an error, so went through and replaced the "/" character with
os.sep everywhere ink is using the open command on a file.